All bugs fixad, and release version 2 in published

+++ app/src/main/java/com/example/quotes/adapters/OwnCursorCustomAdapter.kt
@@ -1,7 +1,10 @@
 package com.example.quotes.adapters
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.database.Cursor
+import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -25,13 +28,29 @@ class OwnCursorCustomAdapter(context: Context, cursor: Cursor) : CursorAdapter(c
         val quoteTextView: TextView? = view?.findViewById(R.id.quote_text)
         val quoteTextAuthor: TextView? = view?.findViewById(R.id.author_text)
         val cardView: CardView? = view?.findViewById(R.id.card_view_front)
-        cardView?.setBackgroundResource(R.drawable.categories_background)
+        cardView?.setBackgroundResource(R.drawable.favorite_card_background)
 
         quoteTextView?.text = cursor?.getString(1)
         quoteTextAuthor?.text = cursor?.getString(2)
         cardView?.setOnLongClickListener {
-            Snackbar.make(view, "Quote deleted. On next refresh, quote will be gone.", Snackbar.LENGTH_SHORT).show()
-            database.removeFromTableOwn(id!!.toInt())
+            val builder = AlertDialog.Builder(view.context).setTitle("Delete").setMessage(
+                "Are you sure you want to delete this Quote?"
+            ).setPositiveButton(R.string.yes_choice) { dialog, which ->
+                Snackbar.make(view, "Quote deleted.", Snackbar.LENGTH_SHORT).show()
+                database.removeFromTableOwn(id!!.toInt())
+                swapCursor(SampleSQLiteDBHelper(view.context).readFromDBTableOwn())
+                dialog.dismiss()
+            }.setNegativeButton(R.string.no_choice) { dialogInterface: DialogInterface, i: Int ->
+                dialogInterface.cancel()
+            }
+            val alertDialog = builder.create()
+            alertDialog.show()
+            val buttonPositive = alertDialog.getButton(DialogInterface.BUTTON_POSITIVE)
+            val buttonNegative = alertDialog.getButton(DialogInterface.BUTTON_NEGATIVE)
+
+            buttonNegative.setTextColor(Color.BLACK)
+            buttonPositive.setTextColor(Color.BLACK)
+
             true
         }
     }
